requirements.txt
anthropic>=0.40.0
beautifulsoup4>=4.12.0
html2text>=2024.2.26
playwright>=1.40.0
pyyaml>=6.0
requests>=2.31.0
